requirements.txt
psycopg2-binary
pandas

apache-airflow-providers-amazon
minio
boto3